backend/reports/requirements.txt
nameko==3.0.0-rc11
python-dotenv
pytest
pytest-cov
langchain
langchain-core
langchain-google-genai
sqlmodel
sqlmodel-controller
